WIP: Move backup generation to serverside
